+++ jac64-core/src/main/java/com/dreamfabric/jac64/emu/bus/AddressableIf.java
@@ -2,15 +2,15 @@ package com.dreamfabric.jac64.emu.bus;
 
 public interface AddressableIf {
 
-    public boolean write(int address, byte data);
+    public boolean write(int address, int data);
 
-    public Byte read(int address);
-    
+    public Integer read(int address);
+
     public int getStartAddress();
-    
+
     public int getEndAddress();
-    
+
     public void setEnabled(boolean enabled);
-    
+
     public boolean isEnabled();
 }
